+++ src/Converter/ConverterFactory.php
@@ -224,14 +224,15 @@ class ConverterFactory
 
             if (is_dir($path)) {
                 // Recursively find files in this directory
-                $result = array_merge($result, $this->findClassNames($path, $fileName));
+                $newBaseDirectory = $baseDirectory ? $baseDirectory . '/' . $fileName : $fileName;
+                $result = array_merge($result, $this->findClassNames($path, $newBaseDirectory));
             } else {
                 // Remove the extension
                 $fileName = pathinfo($fileName, PATHINFO_FILENAME);
 
                 // Get the class name
                 $className = 'Smile\GdprDump\Converter\\';
-                $className .= $baseDirectory ? $baseDirectory . '\\' . $fileName : $fileName;
+                $className .= $baseDirectory ? str_replace('/', '\\', $baseDirectory) . '\\' . $fileName : $fileName;
 
                 // Include only classes that implement the converter interface
                 $reflection = new ReflectionClass($className);